创建了文件系统这个文件夹，计划通过正则表达式做成递归生成name_list的功能

+++ src/base/move_controller.rs
@@ -1,12 +1,12 @@
 use crate::base::facing::Facing;
 pub struct MoveController {
     move_state: MoveState,
-    speed: Option<i32>
+    speed: Option<f32>
 }
 
 //构造函数的集合
 impl MoveController {
-    pub fn new(speed:Option<i32>) -> Self {
+    pub fn new(speed:Option<f32>) -> Self {
         MoveController {
             move_state: MoveState::Stop,
             speed
